# Conflicts:
#	jbb-installation-assets/src/main/resources/liquibase/jbb-db-changelog-0.9.0.xml

+++ domain-services/jbb-frontend/src/main/java/org/jbb/frontend/impl/faq/logic/FaqServiceImpl.java
@@ -12,12 +12,16 @@ package org.jbb.frontend.impl.faq.logic;
 
 import com.google.common.collect.Lists;
 import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
+import javax.validation.ConstraintViolation;
+import javax.validation.Validator;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.lang3.Validate;
 import org.jbb.frontend.api.faq.Faq;
 import org.jbb.frontend.api.faq.FaqCategory;
 import org.jbb.frontend.api.faq.FaqEntry;
+import org.jbb.frontend.api.faq.FaqException;
 import org.jbb.frontend.api.faq.FaqService;
 import org.jbb.frontend.impl.faq.dao.FaqCategoryRepository;
 import org.jbb.frontend.impl.faq.model.FaqCategoryEntity;
@@ -30,22 +34,29 @@ import org.springframework.transaction.annotation.Transactional;
 public class FaqServiceImpl implements FaqService {
 
     private final FaqCategoryRepository faqCategoryRepository;
+    private final Validator validator;
 
     @Override
     public Faq getFaq() {
         List<FaqCategory> faqCategories = faqCategoryRepository.findByOrderByPosition().stream()
             .map(entity -> (FaqCategory) entity)
             .collect(Collectors.toList());
-        return Faq.builder().faqCategories(faqCategories).build();
+        return Faq.builder().categories(faqCategories).build();
     }
 
     @Override
     @Transactional
     public void setFaq(Faq faq) {
         Validate.notNull(faq);
+
+        Set<ConstraintViolation<Faq>> constraintViolations = validator.validate(faq);
+        if (!constraintViolations.isEmpty()) {
+            throw new FaqException(constraintViolations);
+        }
+
         faqCategoryRepository.deleteAll();
 
-        List<FaqCategory> faqCategories = faq.getFaqCategories();
+        List<FaqCategory> faqCategories = faq.getCategories();
         int categoryCount = faqCategories.size();
 
         for (int i = 1; i <= categoryCount; i++) {
